% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tu_maps.R
\name{tu_maps}
\alias{tu_maps}
\title{Time Use Totals for Specific Activity Across US States in R}
\usage{
tu_maps(df, activity, path_to_activity_codes = "data/ATUS_codes.csv", w = NULL)
}
\arguments{
\item{df}{an ATUS-X dataframe constructed by ipumsr package.}

\item{activity}{activity of interest. The name should appear in the .csv file mentioned in the path_to_activity_codes.}

\item{path_to_activity_codes}{a path to an .csv file with the complete coding of activities. An example of the file can be found
here: https://github.com/Kolpashnikova/Kolpashnikova.github.io/blob/master/data/ATUS_codes.csv}

\item{w}{the name of the weights variables in the ATUS-X extract. The default is NULL. Usually, the weights variable is called "WT06."}
}
\description{
Creates a dataframe with 51 variables, corresponding to the US states and DC (two-letter lowercase codes).
Its values represent the mean for the activity of interest for a given state.
}
\references{
Kolpashnikova, Kamila. (2022). Time Use Package for R. Toronto,ON: York University.
}
