% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tu_path.R
\name{tu_path}
\alias{tu_path}
\title{Time Use Path Files in R}
\usage{
tu_path(
  df,
  path_to_activity_codes = "data/ATUS_codes.csv",
  filename1 = "ex_diaries.txt",
  filename2 = "activities.txt",
  filename3 = "weights.txt",
  w = NULL
)
}
\arguments{
\item{df}{an ATUS-X dataframe constructed by ipumsr package.}

\item{path_to_activity_codes}{a path to an .csv file with the complete coding of activities. An example of the file can be found
here: https://github.com/Kolpashnikova/Kolpashnikova.github.io/blob/master/data/ATUS_codes.csv}

\item{filename1}{the path and name of the file where to save the long format diaries.}

\item{filename2}{the path and name of the file where to save the activities dictionary.}

\item{filename3}{the path and name of the file where to save weights.}

\item{w}{the name of the weights variables in the ATUS-X extract. The default is NULL. Usually, the weights variable is called "WT06."}
}
\description{
Create 3 files. Filename1 is for the diaries in the long format for each minute.
Filename2 contains a JSON dictionary, where keys are activity names and values
are the numbers they correspond to. Filename3 contains weights for all the observations
in the same order as in Filename1.
}
\references{
Kolpashnikova, Kamila. (2022). Time Use Package for R. Toronto,ON: York University.
}
