% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tu_totals.R
\name{tu_totals}
\alias{tu_totals}
\title{Time Use Totals in R}
\usage{
tu_totals(df, path_to_activity_codes = "data/ATUS_codes.csv")
}
\arguments{
\item{df}{an ATUS-X dataframe constructed by ipumsr package.}

\item{path_to_activity_codes}{a path to an .csv file with the complete coding of activities. An example of the file can be found
here: https://github.com/Kolpashnikova/Kolpashnikova.github.io/blob/master/data/ATUS_codes.csv}
}
\description{
Creates a list of two variables, "totals" and "activities." The "totals" variable represents
a list of n (where n is the number of coded activities) of percent this activity takes up
in the total of all activities across all observations. The "activities" variable
represents a character list of coded activities in the same order as in they appear in the
totals.
}
\references{
Kolpashnikova, Kamila. (2022). Time Use Package for R. Toronto,ON: York University.
}
