% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tu_transitions.R
\name{tu_transitions}
\alias{tu_transitions}
\title{Time Use Transitions in R}
\usage{
tu_transitions(df, path_to_activity_codes = "data/ATUS_codes.csv", w = NULL)
}
\arguments{
\item{df}{an ATUS-X dataframe constructed by ipumsr package.}

\item{path_to_activity_codes}{a path to an .csv file with the complete coding of activities. An example of the file can be found
here: https://github.com/Kolpashnikova/Kolpashnikova.github.io/blob/master/data/ATUS_codes.csv}

\item{w}{the name of the weights variables in the ATUS-X extract. The default is NULL. Usually, the weights variable is called "WT06."}
}
\description{
Creates a list of two variables, "trate" and "activities." The "trate" variable
contains a n x n matrix (where n is the number of coded activities). Rows represent
the number of the activity from which the transition started, and columns represent
activities into which the transition ended. Transitions of an activity into itself is
always zero. The numbers in the matrix represent the percent of the total transitions
(sum of matrix). The "activities" variable contains a character list of activities in
the same order as they appear as the rows and columns in the trate matrix.
}
\references{
Kolpashnikova, Kamila. (2022). Time Use Package for R. Toronto,ON: York University.
}
